Refactor to account for iframe form/complex elements, add ifReady method to account for onload/dynamic load iframes
Resolve issue with previous update preventing form or complex elements requiring input from being functional.  This was due to the window.focus() triggering after every window blur event, which works fine for simple elements like buttons or things you just click or scroll within, but form inputs become unable to accept input within the iframe since it keeps losing focus.

This update lets you designate an element as a form in the track function so it knows not to apply window.focus() after a blur on it.

This update also adds ifReady, a wrapper function that can be used to wait for an iframe to load before tracking, either being dynamically generated or already in the embed code already in the dom. Track config options all moved into a config obj as the 3rd arg.