add interactions condition to untrack logic in case once = true
If once is set to true in the config and start interactions is past that, the previous code will move the tracked element to untracked even though the callback hasn't been run yet. Thus, the callback never gets run if once is set to true. 

This update adds a condition in updateState to also check interactions vs start when checking the once flag.